+++ Domain.Pagos/Factories/IPagoFactory.cs
@@ -9,7 +9,7 @@ namespace Domain.Pagos.Factories
 {
     public interface IPagoFactory
     {
-        Pago CrearPago(Guid idPago, Guid proyectoId, float monto, int estado);
+        Pago CrearPago(Guid idPago, Guid proyectoId, decimal monto, int estado);
         Pago CrearPago();
     }
 }
